feat: refactor export excel file logic... please don't break...
