+++ src/components/SubstrateAndInvertEditor.tsx
@@ -1,7 +1,7 @@
-import React, { useRef, useState } from 'react';
+import React, { useState } from 'react';
 import { HotTable } from '@handsontable/react';
 import 'handsontable/dist/handsontable.full.min.css';
-import ExcelJS, { Workbook } from 'exceljs';
+import ExcelJS from 'exceljs';
 import {
   CheckboxCellType,
   NumericCellType,
@@ -18,9 +18,15 @@ import {
   MergeCells,
   registerPlugin,
 } from 'handsontable/plugins';
-import { fetchTemplateFromBlobStorage } from '@/utils/azureBlobStorageHelper';
-import { getExcelTemplateSasTokenCookie } from '@/services/excelTemplateSasTokenApi';
 import useSnackbarStore from '@/stores/snackbarStore';
+import {
+  extractApiDataFromWorksheet,
+  getExcelTemplateFiles,
+  getSlateConfig,
+  handleUpdateExcel,
+  readBlobAsArrayBuffer,
+} from '@/utils/exportExcelHelper';
+import { SlateState } from '@/stores/types';
 
 registerCellType(CheckboxCellType);
 registerCellType(NumericCellType);
@@ -35,112 +41,11 @@ registerPlugin(HiddenRows);
 registerPlugin(MergeCells);
 
 interface ExportEditorProps {
-  type: 'substrate' | 'fishInverts';
-  excelBlobData?: Blob | File | null;
-}
-
-// Stuff to set, probably the settings are wrong too cause I suck
-const OcrResultsConfig: Record<string, SlateConfig.SlateConfig> = Object.freeze(
-  {
-    substrate: {
-      rowNumberFrom: 10,
-      rowNumberTo: 31,
-      numberOfColumns: 16,
-      ignoredRows: [],
-      ignoredCols: [],
-    },
-    fishInverts: {
-      rowNumberFrom: 9,
-      rowNumberTo: 64,
-      numberOfColumns: 5,
-      ignoredRows: [24, 25, 26, 45, 48],
-      ignoredCols: [1],
-    },
-  },
-);
-
-// Stuff to set, probably the settings are wrong too cause I suck
-const SlateTypeConfig: Record<string, SlateConfig.SlateConfig> = Object.freeze({
-  substrate: {
-    rowNumberFrom: 13,
-    rowNumberTo: 34,
-    numberOfColumns: 16,
-    ignoredRows: [],
-    ignoredCols: [],
-  },
-  fishInverts: {
-    rowNumberFrom: 9,
-    rowNumberTo: 64,
-    numberOfColumns: 5,
-    ignoredRows: [24, 25, 26, 45, 48],
-    ignoredCols: [1],
-  },
-});
-
-// Read from template or ocrUpload config
-function getSlateConfig(
-  type: 'substrate' | 'fishInverts',
-  isOcrResultsConfig?: boolean,
-) {
-  return isOcrResultsConfig ? OcrResultsConfig[type] : SlateTypeConfig[type];
-}
-
-function extractApiDataFromWorksheet(
-  workbook: Workbook,
-  config: SlateConfig.SlateConfig,
-): { extractedData: (string | number)[][]; styles: any } {
-  const worksheet = workbook.worksheets[0]; // Assuming the first sheet
-  const jsonData: (string | number)[][] = [];
-  const styles: any = {};
-
-  worksheet.eachRow({ includeEmpty: true }, (row, rowNumber) => {
-    if (rowNumber >= config.rowNumberFrom && rowNumber <= config.rowNumberTo) {
-      if (config.ignoredRows.includes(rowNumber)) {
-        jsonData.push(Array(config.numberOfColumns).fill(null)); // Placeholder for ignored rows
-      } else {
-        const rowValues = row.values as (string | number)[];
-        rowValues.shift(); // Remove the first element which is usually the row number itself
-
-        // Capture styles (LOGIC HERE PROBABLY WRONG, THE DISPLAY IN THE EXCEL REALLY NEED A FIX)
-        row.eachCell({ includeEmpty: true }, (cell, colNumber) => {
-          if (!styles[rowNumber - 1]) styles[rowNumber - 1] = {};
-          styles[rowNumber - 1][colNumber - 1] = cell.style;
-        });
-
-        jsonData.push(rowValues.slice(0, config.numberOfColumns));
-      }
-    }
-  });
-
-  return { extractedData: jsonData, styles };
-}
-// pass worksheet you want to update, data to update, templateConfig, and cellStyles
-function updateWorksheet(
-  worksheet: ExcelJS.Worksheet,
-  data: (string | number)[][],
-  templateConfig: SlateConfig.SlateConfig,
-  // apiConfig: SlateConfig.SlateConfig,
-  cellStyles: any[],
-) {
-  data.forEach((row, rowIndex) => {
-    const actualRowIndex = rowIndex + templateConfig.rowNumberFrom;
-
-    if (templateConfig.ignoredRows.includes(actualRowIndex)) return; // Skip the ignored row
-
-    row.forEach((cell, colIndex) => {
-      if (templateConfig.ignoredCols.includes(colIndex + 1)) return; // Skip the ignored column
-
-      const cellRef = worksheet.getCell(actualRowIndex, colIndex + 1);
-      cellRef.value = cell;
-
-      if (cellStyles[rowIndex] && cellStyles[rowIndex][colIndex]) {
-        cellRef.style = cellStyles[rowIndex][colIndex];
-      }
-    });
-  });
+  slate: SlateState;
 }
 
 export default function SubstrateAndInvertEditor(props: ExportEditorProps) {
+  const { slate } = props;
   const [exportFileData, setExportFileData] = useState<(string | number)[][]>(
     [],
   );
@@ -149,74 +54,7 @@ export default function SubstrateAndInvertEditor(props: ExportEditorProps) {
   const [cellStyles, setCellStyles] = useState<any>({});
   const addMessage = useSnackbarStore((state) => state.addMessage);
 
-  const templateConfig: SlateConfig.SlateConfig = getSlateConfig(props.type);
-  // const config: SlateConfig.SlateConfig = getSlateConfig(props.type);
-
-  // I might be doing more than just "getExcelTemplatefiles"
-  const getExcelTemplateFiles = async () => {
-    try {
-      const sasToken = await getExcelTemplateSasTokenCookie();
-
-      // console.log('SasToken generated successfully: ', sasToken);
-
-      const blobFromStorage = await fetchTemplateFromBlobStorage(
-        sasToken.value,
-        props.type,
-      );
-
-      if (props.excelBlobData) {
-        // console.log('Excel Blob Data: ', props.excelBlobData);
-
-        const apiArrayBuffer = await readBlobAsArrayBuffer(props.excelBlobData);
-        const apiWorkbook = new ExcelJS.Workbook();
-        await apiWorkbook.xlsx.load(apiArrayBuffer);
-
-        const apiConfig = getSlateConfig(props.type, true);
-        const { extractedData: apiData, styles: apiStyles } =
-          extractApiDataFromWorksheet(apiWorkbook, apiConfig);
-
-        const templateArrayBuffer =
-          await readBlobAsArrayBuffer(blobFromStorage);
-        const templateWorkbook = new ExcelJS.Workbook();
-        await templateWorkbook.xlsx.load(templateArrayBuffer);
-
-        const templateWorksheet = templateWorkbook.worksheets[0];
-        updateWorksheet(
-          templateWorksheet,
-          apiData,
-          templateConfig,
-          //   apiConfig,
-          apiStyles,
-        );
-
-        const buffer = await templateWorkbook.xlsx.writeBuffer();
-        const updatedBlob = new Blob([buffer], {
-          type: 'application/octet-stream',
-        });
-        setBlobData(updatedBlob);
-
-        parseBlobData(updatedBlob);
-      }
-    } catch (error: any) {
-      addMessage(error.message, 'error');
-    }
-  };
-
-  // Helper function to convert readable stream to buffer
-  async function readBlobAsArrayBuffer(blob: Blob): Promise<ArrayBuffer> {
-    return new Promise<ArrayBuffer>((resolve, reject) => {
-      const reader = new FileReader();
-      reader.onload = () => {
-        if (reader.result instanceof ArrayBuffer) resolve(reader.result);
-        else reject(new Error('Failed to read blob as ArrayBuffer'));
-      };
-
-      reader.onerror = () => {
-        reject(reader.error);
-      };
-      reader.readAsArrayBuffer(blob);
-    });
-  }
+  const templateConfig: SlateConfig.SlateConfig = getSlateConfig(slate.type);
 
   // PARSE data to be used in the handsontable
   async function parseBlobData(blob: Blob) {
@@ -239,87 +77,26 @@ export default function SubstrateAndInvertEditor(props: ExportEditorProps) {
     }
   }
 
-  // File Upload function that was previously used. Leaving it here because it was hard work ;_;
-  const handleFileUpload = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
+  React.useEffect(() => {
+    const fetchExcelTemplateFiles = async () => {
+      try {
+        const updatedBlob = await getExcelTemplateFiles(
+          props.slate,
+          templateConfig,
+        );
 
-    const reader = new FileReader();
-
-    reader.onload = async (event) => {
-      const arrayBuffer = event.target?.result as ArrayBuffer;
-      const workbook = new ExcelJS.Workbook();
-      await workbook.xlsx.load(arrayBuffer);
-
-      const worksheet = workbook.worksheets[0]; // Assuming the first sheet
-      const jsonData: (string | number)[][] = [];
-      const styles: any = {};
-
-      worksheet.eachRow({ includeEmpty: true }, (row, rowNumber) => {
-        // Extract rows 13 to 34
-        if (rowNumber >= 13 && rowNumber <= 34) {
-          const rowValues = row.values as (string | number)[];
-          rowValues.shift(); // Remove the first element which is usually the row number itself
-
-          // Capture styles ( the logic here should probably be wrong, THIS NEEDS REFACTORING FOR EXCEL DISPLAY TO LOOK GOOD)
-          row.eachCell({ includeEmpty: true }, (cell, colNumber) => {
-            if (!styles[rowNumber - 1]) styles[rowNumber - 1] = {};
-            styles[rowNumber - 1][colNumber - 1] = cell.style;
-          });
-
-          jsonData.push(rowValues.slice(0, 16));
+        if (updatedBlob) {
+          setBlobData(updatedBlob);
+          parseBlobData(updatedBlob);
+        } else {
+          setBlobData(null);
         }
-      });
-
-      setExportFileData(jsonData);
-      setCellStyles(styles);
+      } catch (error: any) {
+        addMessage(error.message, 'error');
+      }
     };
 
-    reader.readAsArrayBuffer(file);
-  };
-
-  const handleUpdateExcel = async () => {
-    if (!blobData) {
-      console.log('Currently no file selected.');
-      return;
-    }
-
-    //worksheet will be taken and looped through template Config
-    const arrayBuffer = await readBlobAsArrayBuffer(blobData);
-    const workbook = new ExcelJS.Workbook();
-    await workbook.xlsx.load(arrayBuffer);
-
-    const worksheet = workbook.worksheets[0];
-
-    const updatedExcelConfig = templateConfig;
-    const updatedCellStyles = cellStyles;
-
-    // Update the worksheet with data and preserve styles
-    // passing it templateConfig
-    updateWorksheet(
-      worksheet,
-      exportFileData,
-      updatedExcelConfig,
-      updatedCellStyles,
-    );
-
-    const buffer = await workbook.xlsx.writeBuffer();
-    const updatedBlob = new Blob([buffer], {
-      type: 'application/octet-stream',
-    });
-
-    const url = window.URL.createObjectURL(updatedBlob);
-    const a = document.createElement('a');
-    a.href = url;
-    a.download = 'updated_file.xlsx';
-    document.body.appendChild(a);
-    a.click();
-    document.body.removeChild(a);
-    window.URL.revokeObjectURL(url);
-  };
-
-  React.useEffect(() => {
-    getExcelTemplateFiles();
+    fetchExcelTemplateFiles();
   }, []);
 
   return (
@@ -334,7 +111,18 @@ export default function SubstrateAndInvertEditor(props: ExportEditorProps) {
             height="auto"
             licenseKey="non-commercial-and-evaluation"
           />
-          <button onClick={handleUpdateExcel}>Update Excel File</button>
+          <button
+            onClick={() =>
+              handleUpdateExcel(
+                blobData,
+                templateConfig,
+                cellStyles,
+                exportFileData,
+              )
+            }
+          >
+            Update Excel File
+          </button>
         </div>
       )}
     </div>
